[Modify]: elec skin adapter for multiple definition
1.适配include的时候多重包含导致的问题
