+++ elec_skin/include/elec_skin/elec_skin.hpp
@@ -21,6 +21,7 @@
 #include "ament_index_cpp/get_package_share_directory.hpp"
 #include "embed_protocol/embed_protocol.hpp"
 #include "cyberdog_common/cyberdog_log.hpp"
+#include "elec_skin/elec_skin_base.hpp"
 
 #define EVM cyberdog::embed
 
@@ -37,40 +38,7 @@ typedef struct _can_data
   uint8_t data3;
 } can_data;
 
-enum class ModelSwitch : uint8_t
-{
-  MS_FLASH = 0,    // 闪烁
-  MS_WAVEF = 1,    // 动画前向后变
-  MS_RANDOM = 2,   // 随机
-  MS_WAVEB = 3,    // 动画后向前变
-  MS_CONTROL = 4,  // 上位机实时控制模式
-};  // enum class PositionSkin
-
-enum class PositionSkin : uint8_t
-{
-  PS_BODYM = 0,  // 背部
-  PS_LBLEG = 1,  // 左后腿
-  PS_BODYL = 2,  // 左侧
-  PS_LFLEG = 3,  // 左前腿
-  PS_FRONT = 4,  // 前胸
-  PS_RFLEG = 5,  // 右前腿
-  PS_BODYR = 6,  // 右侧
-  PS_RBLEG = 7,  // 右后腿
-};  // enum class PositionSkin
-
-enum class PositionColorChangeDirection : uint8_t
-{
-  PCCD_WTOB = 0,   // 白变黑
-  PCCD_BTOW = 1,   // 黑变白
-};  // enum class PositionColorChangeDirection
-
-enum class PositionColorStartDirection : uint8_t
-{
-  PCSD_FRONT = 0,   // 前方先变色
-  PCSD_BACK  = 1,   // 后方先变色
-};  // enum class PositionColorStartDirection
-
-class ElecSkin final
+class ElecSkin final : public ElecSkinBase
 {
 public:
   ElecSkin()
